@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "7.0.3",
+  "version": "7.0.4",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -28,10 +28,10 @@
     "sift": "16.0.1"
   },
   "devDependencies": {
-    "@babel/core": "7.21.0",
-    "@babel/preset-env": "7.20.2",
-    "@typescript-eslint/eslint-plugin": "5.54.0",
-    "@typescript-eslint/parser": "5.54.0",
+    "@babel/core": "7.21.4",
+    "@babel/preset-env": "7.21.4",
+    "@typescript-eslint/eslint-plugin": "5.57.0",
+    "@typescript-eslint/parser": "5.57.0",
     "acquit": "1.3.0",
     "acquit-ignore": "0.2.1",
     "acquit-require": "0.1.1",
@@ -45,28 +45,29 @@
     "crypto-browserify": "3.12.0",
     "dotenv": "16.0.3",
     "dox": "1.0.0",
-    "eslint": "8.35.0",
+    "eslint": "8.37.0",
     "eslint-plugin-markdown": "^3.0.0",
     "eslint-plugin-mocha-no-only": "1.1.1",
     "express": "^4.18.1",
     "highlight.js": "11.7.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "marked": "4.2.12",
+    "marked": "4.3.0",
     "mkdirp": "^2.1.3",
     "mocha": "10.2.0",
     "moment": "2.x",
-    "mongodb-memory-server": "8.11.5",
+    "mongodb-memory-server": "8.12.1",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.2",
     "q": "1.5.1",
-    "sinon": "15.0.1",
+    "sinon": "15.0.3",
     "stream-browserify": "3.0.0",
     "tsd": "0.25.0",
-    "typescript": "4.9.5",
+    "typescript": "5.0.3",
     "uuid": "9.0.0",
-    "webpack": "5.75.0"
+    "webpack": "5.77.0",
+    "fs-extra": "~11.1.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
